+++ internal/models/plan_data.go
@@ -18,9 +18,29 @@
 
 package models
 
-type PlanDataRaw []struct {
+import "fmt"
+
+type PlanData struct {
 	PlanID         int    `json:"plan_id"`
 	PlanName       string `json:"plan_name"`
 	NbLicencesUsed int    `json:"nb_licences_used"`
 	NbLicences     int    `json:"nb_licences"`
+}
+
+func (pd PlanData) TableHeader() []string {
+	return []string{
+		"Plan ID",
+		"Plan Name",
+		"Number of Licences Used",
+		"Total Number of Licences",
+	}
+}
+
+func (pd PlanData) TableRow() []string {
+	return []string{
+		fmt.Sprintf("%d", pd.PlanID),
+		pd.PlanName,
+		fmt.Sprintf("%d", pd.NbLicencesUsed),
+		fmt.Sprintf("%d", pd.NbLicences),
+	}
 }